refactor: change deployment file name
